move all fuzz stuff to _test files
